+++ tse_analytics/core/data/pipeline/time_intervals_binning_pipe_operator.py
@@ -1,7 +1,7 @@
 import pandas as pd
 
 from tse_analytics.core.data.binning import TimeIntervalsBinningSettings
-from tse_analytics.core.data.shared import SplitMode, Variable
+from tse_analytics.core.data.shared import Variable
 
 default_columns = ["DateTime", "Timedelta", "Animal", "Box", "Run", "Bin"]
 
@@ -10,60 +10,42 @@ def process_time_interval_binning(
     df: pd.DataFrame,
     settings: TimeIntervalsBinningSettings,
     variables: dict[str, Variable],
-    split_mode: SplitMode,
-    factor_names: list[str],
-    selected_factor_name: str,
 ) -> pd.DataFrame:
     if df.empty:
         return df
 
-    match split_mode:
-        case SplitMode.ANIMAL:
-            agg = {
-                "DateTime": "first",
-                "Box": "first",
-            }
-            for factor_name in factor_names:
-                agg[factor_name] = "first"
-            group_by = ["Animal"]
-            sort_by = ["Timedelta", "Box"]
-            grouped = df.groupby(group_by, dropna=False, observed=False)
-        case SplitMode.FACTOR:
-            agg = {
-                "DateTime": "first",
-            }
-            group_by = [selected_factor_name]
-            sort_by = ["Timedelta", selected_factor_name]
-            grouped = df.groupby(group_by, dropna=False, observed=False)
-        case SplitMode.RUN:
-            agg = {
-                "DateTime": "first",
-            }
-            group_by = ["Run"]
-            sort_by = ["Timedelta", "Run"]
-            grouped = df.groupby(group_by, dropna=False, observed=False)
-        case _:  # Total split mode
-            agg = {
-                "DateTime": "first",
-            }
-            sort_by = ["Timedelta"]
-            grouped = df
-
-    for variable in variables.values():
-        agg[variable.name] = variable.aggregation
-
     timedelta = pd.Timedelta(f"{settings.delta}{settings.unit}")
-    result = grouped.resample(timedelta, on="Timedelta", origin="start").agg(agg)
 
-    result.sort_values(by=sort_by, inplace=True)
+    agg = {
+        "DateTime": "first",
+        "Box": "first",
+        "Run": "first",
+    }
+    for column in df.columns:
+        if column not in default_columns:
+            if df.dtypes[column].name != "category":
+                agg[column] = variables[column].aggregation
+            else:
+                agg[column] = "first"
 
-    # the inverse of groupby, reset_index
-    result.reset_index(inplace=True)
+    result = df.groupby("Animal", dropna=False, observed=False)
+    result = result.resample(timedelta, on="Timedelta", origin="start").aggregate(agg)
+    result.sort_values(by="DateTime", inplace=True)
+    result.reset_index(inplace=True, drop=False)
 
-    # start_date_time = result["DateTime"].iloc[0]
-    # result["Timedelta"] = result["DateTime"] - start_date_time
+    # Get unique runs numbers
+    runs = result["Run"].unique().tolist()
 
+    # Reassign timedeltas
+    for run in runs:
+        # Get start timestamp per run
+        start_date_time = result[result["Run"] == run]["DateTime"].iloc[0]
+        result.loc[result["Run"] == run, "Timedelta"] = result["DateTime"] - start_date_time
+
+    # Reassign bins numbers
     result["Bin"] = (result["Timedelta"] / timedelta).round().astype(int)
-    # result = result.astype({"Bin": "category"})
+
+    result.sort_values(by="Bin", inplace=True)
+    result.reset_index(inplace=True, drop=True)
 
     return result