+++ tse_analytics/core/data/pipeline/time_phases_binning_pipe_operator.py
@@ -1,18 +1,15 @@
 import pandas as pd
 
 from tse_analytics.core.data.binning import TimePhasesBinningSettings
-from tse_analytics.core.data.shared import SplitMode, Variable
+from tse_analytics.core.data.shared import Variable
 
-default_columns = ["Timedelta", "Box", "Run"]
+default_columns = ["Timedelta", "Animal", "Box", "Run", "Bin"]
 
 
 def process_time_phases_binning(
     df: pd.DataFrame,
     settings: TimePhasesBinningSettings,
     variables: dict[str, Variable],
-    split_mode: SplitMode,
-    factor_names: list[str],
-    selected_factor_name: str,
 ) -> pd.DataFrame:
     settings.time_phases.sort(key=lambda x: x.start_timestamp)
 
@@ -27,30 +24,37 @@ def process_time_phases_binning(
     categories = [item.name for item in settings.time_phases]
     df["Bin"] = df["Bin"].cat.set_categories(categories, ordered=True)
 
-    match split_mode:
-        case SplitMode.ANIMAL:
-            agg = {
-                "Box": "first",
-            }
-            for factor_name in factor_names:
-                agg[factor_name] = "first"
-            group_by = ["Animal", "Bin"]
-        case SplitMode.FACTOR:
-            agg = {}
-            group_by = [selected_factor_name, "Bin"]
-        case SplitMode.RUN:
-            agg = {}
-            group_by = ["Run", "Bin"]
-        case _:
-            agg = {}
-            group_by = ["Bin"]
+    # match split_mode:
+    #     case SplitMode.ANIMAL:
+    #         agg = {
+    #             "Box": "first",
+    #         }
+    #         for factor_name in factor_names:
+    #             agg[factor_name] = "first"
+    #         group_by = ["Animal", "Bin"]
+    #     case SplitMode.FACTOR:
+    #         agg = {}
+    #         group_by = [selected_factor_name, "Bin"]
+    #     case SplitMode.RUN:
+    #         agg = {}
+    #         group_by = ["Run", "Bin"]
+    #     case _:
+    #         agg = {}
+    #         group_by = ["Bin"]
 
-    for variable in variables.values():
-        agg[variable.name] = variable.aggregation
+    agg = {}
+    for column in df.columns:
+        if column not in default_columns:
+            if df.dtypes[column].name != "category":
+                agg[column] = variables[column].aggregation
+            else:
+                agg[column] = "first"
 
-    result = df.groupby(group_by, dropna=False, observed=False).agg(agg)
+    if len(agg) == 0:
+        return df
 
-    # the inverse of groupby, reset_index
+    result = df.groupby(["Animal", "Bin"], dropna=False, observed=False).aggregate(agg)
+    # result.sort_values(by=["Animal", "Bin"], inplace=True)
     result.reset_index(inplace=True)
 
     return result