+++ tse_analytics/views/analysis/dimensionality_widget.py
@@ -9,6 +9,7 @@ from PySide6.QtWidgets import QAbstractItemView, QWidget
 from sklearn.decomposition import PCA
 from sklearn.manifold import TSNE
 
+from tse_analytics.core.data.binning import BinningMode
 from tse_analytics.core.data.shared import SplitMode, Variable
 from tse_analytics.core.helper import make_toast, show_help
 from tse_analytics.core.manager import Manager
@@ -75,6 +76,24 @@ class DimensionalityWidget(QWidget, MessengerListener):
             self.ui.factorSelector.clear()
 
     def _update(self):
+        if (
+            Manager.data.selected_dataset.binning_settings.apply
+            and (
+                Manager.data.selected_dataset.binning_settings.mode == BinningMode.PHASES
+                or Manager.data.selected_dataset.binning_settings.mode == BinningMode.CYCLES
+            )
+            and self.ui.radioButtonSplitByRun.isChecked()
+        ):
+            make_toast(
+                self,
+                "Dimensionality Analysis",
+                "Split by Run not available when binning by cycles or time phases is active.",
+                duration=4000,
+                preset=ToastPreset.WARNING,
+                show_duration_bar=True,
+            ).show()
+            return
+
         if self.ui.radioButtonMatrixPlot.isChecked():
             self._update_matrix_plot()
         elif self.ui.radioButtonPCA.isChecked():
@@ -227,7 +246,7 @@ class DimensionalityWidget(QWidget, MessengerListener):
                     "1": "PC 2",
                     "color": by,
                 },
-                hover_name=df["Animal"],
+                # hover_name=df["Animal"],
             )
         elif n_components == 3:
             fig = px.scatter_3d(
@@ -243,7 +262,7 @@ class DimensionalityWidget(QWidget, MessengerListener):
                     "2": "PC 3",
                     "color": by,
                 },
-                hover_name=df["Animal"],
+                # hover_name=df["Animal"],
             )
 
         file = QTemporaryFile(f"{QDir.tempPath()}/XXXXXX.html", self)